backup_repo_function/requirements.txt
GitPython==3.1.0
google-cloud-storage==1.26.0
google-cloud-secret-manager==0.2.0
